move geth to a single file
