+++ docker/test_env/env_component_test.go
@@ -51,7 +51,7 @@ func TestEnvComponentPauseChaos(t *testing.T) {
 		justTimeWithoutMicrosecs := justTime[:len(justTime)-7]
 
 		// blocking
-		err = g.(*GethPoa).ChaosPause(l, 5*time.Second)
+		err = g.(*Geth).ChaosPause(l, 5*time.Second)
 
 		// check that there were no logs when paused
 		for _, lo := range consumer.Msgs {
@@ -63,12 +63,12 @@ func TestEnvComponentPauseChaos(t *testing.T) {
 
 	t.Run("check container traffic can be lost", func(t *testing.T) {
 		// TODO: assert with a busybox container that the traffic is lost
-		err = g.(*GethPoa).ChaosNetworkLoss(l, 30*time.Second, 100, "", nil, nil, nil)
+		err = g.(*Geth).ChaosNetworkLoss(l, 30*time.Second, 100, "", nil, nil, nil)
 		require.NoError(t, err)
 	})
 	t.Run("check container latency can be changed", func(t *testing.T) {
 		// TODO: assert with a busybox container that the traffic is delayed
-		err = g.(*GethPoa).ChaosNetworkDelay(l, 30*time.Second, 5*time.Second, "", nil, nil, nil)
+		err = g.(*Geth).ChaosNetworkDelay(l, 30*time.Second, 5*time.Second, "", nil, nil, nil)
 		require.NoError(t, err)
 	})
 }
