+++ docker/test_env/geth_test.go
@@ -8,9 +8,48 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/smartcontractkit/chainlink-testing-framework/blockchain"
+	"github.com/smartcontractkit/chainlink-testing-framework/docker"
 	"github.com/smartcontractkit/chainlink-testing-framework/logging"
 	"github.com/smartcontractkit/chainlink-testing-framework/utils/testcontext"
 )
+
+func TestOldGeth(t *testing.T) {
+	l := logging.GetTestLogger(t)
+	network, err := docker.CreateNetwork(l)
+	require.NoError(t, err)
+	defaultChainCfg := GetDefaultChainConfig()
+	g := NewGethPoa([]string{network.Name}, &defaultChainCfg).
+		WithTestInstance(t)
+	_, err = g.StartContainer()
+	require.NoError(t, err)
+	ns := blockchain.SimulatedEVMNetwork
+	ns.URLs = []string{g.GetExternalWsUrl()}
+	c, err := blockchain.ConnectEVMClient(ns, l)
+	require.NoError(t, err, "Couldn't connect to the evm client")
+	err = c.Close()
+	require.NoError(t, err, "Couldn't close the client")
+}
+
+func TestEth1WithGeth(t *testing.T) {
+	l := logging.GetTestLogger(t)
+
+	builder := NewEthereumNetworkBuilder()
+	cfg, err := builder.
+		WithConsensusType(ConsensusType_PoW).
+		WithExecutionLayer(ExecutionLayer_Geth).
+		Build()
+	require.NoError(t, err, "Builder validation failed")
+
+	_, eth2, err := cfg.Start()
+	require.NoError(t, err, "Couldn't start PoW network")
+
+	ns := blockchain.SimulatedEVMNetwork
+	ns.URLs = eth2.PublicWsUrls()
+	c, err := blockchain.ConnectEVMClient(ns, l)
+	require.NoError(t, err, "Couldn't connect to the evm client")
+	err = c.Close()
+	require.NoError(t, err, "Couldn't close the client")
+}
 
 func TestEth2WithPrysmAndGeth(t *testing.T) {
 	l := logging.GetTestLogger(t)
